# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study/materials:
    get:
      operationId: listStudyMaterials
      tags: [study-materials]
      summary: List learner's study materials
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterial"
    post:
      operationId: createStudyMaterial
      tags: [study-materials]
      summary: Ingest new study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyMaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterial"
  /study/materials/{materialId}:
    get:
      operationId: getStudyMaterial
      tags: [study-materials]
      summary: Get a single material
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterial"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyError"
    delete:
      operationId: deleteStudyMaterial
      tags: [study-materials]
      summary: Delete a material and its concepts/flashcards
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySuccess"
  /study/materials/{materialId}/concepts:
    get:
      operationId: listStudyConcepts
      tags: [study-concepts]
      summary: Get AI-extracted concepts for a material
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyConcept"

  # ─── Study Flashcards ───
  /study/flashcards:
    get:
      operationId: listStudyFlashcards
      tags: [study-flashcards]
      summary: List all flashcards for the learner
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyFlashcard"
    post:
      operationId: createStudyFlashcard
      tags: [study-flashcards]
      summary: Create a flashcard manually or from a concept
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyFlashcardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyFlashcard"
  /study/flashcards/{flashcardId}:
    patch:
      operationId: updateStudyFlashcard
      tags: [study-flashcards]
      summary: Update a flashcard
      parameters:
        - name: flashcardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyFlashcardUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyFlashcard"
    delete:
      operationId: deleteStudyFlashcard
      tags: [study-flashcards]
      summary: Delete a flashcard
      parameters:
        - name: flashcardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySuccess"
...
  /study/flashcards/{flashcardId}/review:
    post:
      operationId: reviewStudyFlashcard
      tags: [study-flashcards]
      summary: Submit a review result (SM-2)
      parameters:
        - name: flashcardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyReviewInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyFlashcard"

  # ─── Study Practice ───
...
components:
  schemas:
...
    StudyError:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required:
        - error
    StudySuccess:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    StudyMaterial:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        sourceType:
          type: string
        sourceUrl:
          type: ["string", "null"]
        contentText:
          type: string
        conceptCount:
          type: integer
        flashcardCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - sourceType
        - contentText
        - conceptCount
        - flashcardCount
        - createdAt
    StudyMaterialInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        sourceType:
          type: string
          enum: [paste, url, file]
        sourceUrl:
          type: ["string", "null"]
        contentText:
          type: string
          minLength: 1
      required:
        - title
        - sourceType
        - contentText
    StudyConcept:
      type: object
      properties:
        id:
          type: string
        materialId:
          type: string
        title:
          type: string
        explanation:
          type: string
        difficulty:
          type: string
        keyTerms:
          type: array
          items:
            type: string
        relatedConceptIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - materialId
        - title
        - explanation
        - difficulty
        - keyTerms
        - relatedConceptIds
        - createdAt
    StudyFlashcard:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        materialId:
          type: ["string", "null"]
        conceptId:
          type: ["string", "null"]
        front:
          type: string
        back:
          type: string
        hint:
          type: ["string", "null"]
        intervalDays:
          type: number
        repetitions:
          type: integer
        easeFactor:
          type: number
        nextReviewAt:
          type: ["string", "null"]
        lastReviewedAt:
          type: ["string", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - front
        - back
        - intervalDays
        - repetitions
        - easeFactor
        - reviewCount
        - createdAt
    StudyFlashcardInput:
      type: object
      properties:
        materialId:
          type: ["string", "null"]
        conceptId:
          type: ["string", "null"]
        front:
          type: string
          minLength: 1
        back:
          type: string
          minLength: 1
        hint:
          type: ["string", "null"]
      required:
        - front
        - back
    StudyFlashcardUpdate:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        hint:
          type: ["string", "null"]
      required: []
...
    StudyReviewInput:
      type: object
      properties:
        quality:
          type: integer
          minimum: 0
          maximum: 5
          description: SM-2 quality rating (0=complete blackout, 5=perfect)
      required:
        - quality
...